refactor project, implement continent search
